Translator/translator/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-rest-framework==0.1.0
django-restframework==0.0.1
django-rq==2.4.1
django-translation-manager==1.3.0
djangorestframework==3.14.0
greenlet==2.0.1
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.1
polib==1.1.1
psycopg2==2.9.5
pytz==2022.6
redis==4.4.0
requests==2.28.1
rq==1.11.1
SQLAlchemy==1.4.44
sqlparse==0.4.3
translate==3.6.1
tzdata==2022.7
urllib3==1.26.13